# perf: Collect crates in parallel for analysis-stats