The slowest part of analysis-stats is Item Collection for large projects, which processes each crate sequentially. This is slow and crates are independent.

For some of the larger (proprietary) projects I've worked on, this is a significant speedup.

Before:

    $ rust-analyzer analysis-stats .
    Database loaded:     1.23s, 4477minstr, 79mb (metadata 69.45ms, 91minstr, 2634kb; build 952.00ns, 15kinstr, 0b)
      item trees: 113
    Item Tree Collection: 89.27ms, 643minstr, 10mb
      crates: 4, mods: 368, decls: 4258, bodies: 3810, adts: 220, consts: 288
    Item Collection:     86.88s, 337ginstr, 1826mb
    Body lowering:       2.25s, 10ginstr, 82mb
     exprs: 244530, ??ty: 583 (0%), ?ty: 250 (0%), !ty: 16
      pats: 41173, ??ty: 35 (0%), ?ty: 34 (0%), !ty: 0
    Inference:           49.99s, 194ginstr, 739mb
    MIR lowering:        4.51s, 17ginstr, 68mb
    Mir failed bodies: 891 (25%)
    Data layouts:        20.33ms, 78minstr, -1091kb
    Failed data layouts: 26 (12%)
    Const evaluation:    470.00ms, 2544minstr, 4mb
    Failed const evals: 0 (0%)
    Total:               144.35s, 563ginstr, 2731mb

After:

    $ rust-analyzer analysis-stats .
    Database loaded:     1.27s, 4477minstr, 79mb (metadata 69.00ms, 91minstr, 2634kb; build 1.23µs, 15kinstr, 0b)
      item trees: 113
    Item Tree Collection: 91.34ms, 643minstr, 10mb
      crates: 4, mods: 368, decls: 4258, bodies: 3810, adts: 220, consts: 288
    Item Collection:     36.53s, 12minstr, 1890mb
    Body lowering:       1.96s, 10ginstr, 72mb
     exprs: 244530, ??ty: 583 (0%), ?ty: 250 (0%), !ty: 16
      pats: 41173, ??ty: 35 (0%), ?ty: 34 (0%), !ty: 0
    Inference:           47.50s, 182ginstr, 698mb
    MIR lowering:        3.59s, 13ginstr, 66mb
    Mir failed bodies: 891 (25%)
    Data layouts:        17.60ms, 77minstr, 1535kb
    Failed data layouts: 26 (12%)
    Const evaluation:    419.93ms, 2377minstr, 5mb
    Failed const evals: 0 (0%)
    Total:               91.62s, 209ginstr, 2745mb

Item Collection has gone from 86 seconds to 36 seconds here.