@Wilfred 2024-05-09
This should probably respect `self.parallel`. I'll follow up :)